requirements.txt
numpy>=1.24
librosa>=0.10.0 # librosa==0.10.0.post2
pyrubberband==0.3.0
soundfile>=0.12.1
yt_dlp==2023.1.6
sf_segmenter==0.0.2
crepe==0.0.13
demucs==4.0.0
